separate blocking op initialization from thread signal init
